requirements.txt
torch>=2.0.0
transformers>=4.35.0
openai>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
pyyaml>=6.0
